Double count

+++ PoohAPI.Infrastructure.CompanyDB/Models/DBCompany.cs
@@ -27,7 +27,7 @@ namespace PoohAPI.Infrastructure.CompanyDB.Models
         public decimal bedrijf_breedtegraad { get; set; }
         public decimal bedrijf_lengtegraad { get; set; }
         public decimal average_reviews { get; set; }
-        public int amount_reviews { get; set; }
+        public decimal amount_reviews { get; set; }
         public double distance { get; set; }
         public string opleidingen { get; set; }
     }
